make it possible to set the custom-resource-path to any URL
Currently, when the custom-resource-path is set, it always refers the
local filesystem. I'd like to be able to load templates from a jar file
without always specifying the full path. This patch makes it possible to
call set-resource-path! with a jar URL, like in

(set-resource-path! (clojure.java.io/resource "META-INF/foo/templates"))

This currently breaks the related tests and is also missing some tests
for the new functionality.