Fix a bug with elif tags in files.
Using multiple elif clauses in the same if did not
work in file temaplates, only string ones. If you tried it
you got a "no closing tag found" error.

I'm not 100% sure why it worked for strings to begin with..

But anyhow the fix was that the elif tag needed to have itself as a closing
tag.