Optimise if conditions like {% if a.b > a.c %}.
By caching the parsing of the a.b style "filters",
we speed up templates with lots of these quite a lot.

The many-numeric-if-clauses-bench now runs about 3 times faster,
from ~150ms to ~50ms on my machine.

I'm not sure if it could be simplifed a bit, looks a bit complicated now.